structure change. It worked
