+++ stub/orgs/orgs.controller.js
@@ -8,15 +8,6 @@ router.get('/api/organizations', async function getOrgs(req, res, next) {
 
         try {
             const orgsList = await mainService.getOrgs()
-            // convert data from db models to resources
-            // orgs = orgsList.map(function (data) {
-            //     return {
-            //         id: data.id,
-            //         name: data.name,
-            //         district_id: data.district_id
-            //     }
-            // });
-            // console.log(4444)
             return orgsList ? res.json(orgsList) : res.status(404).json({message: 'Организации отсутствуют'})
         } catch (error) {
             next(error)
@@ -26,6 +17,18 @@ router.get('/api/organizations', async function getOrgs(req, res, next) {
     }
 );
 
+router.get('/api/organizations:url', async function getOrgs(req, res, next) {
+
+        try {
+            const data = await mainService.getOrg()
+            return data ? res.json(data) : res.status(404).json({message: 'Организация отсутствуют'})
+        } catch (error) {
+            next(error)
+
+        }
+
+    }
+);
 module.exports = router;
