+++ stub/uds-db.sql
@@ -277,7 +277,7 @@ UNLOCK TABLES;
 
 
 
-/*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
+ /*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
 /*!40101 SET SQL_MODE=@OLD_SQL_MODE */;
 /*!40014 SET FOREIGN_KEY_CHECKS=@OLD_FOREIGN_KEY_CHECKS */;
 /*!40101 SET CHARACTER_SET_CLIENT=@OLD_CHARACTER_SET_CLIENT */;
